+++ src/modules/auth/auth.test.ts
@@ -18,6 +18,7 @@ describe('auth module', () => {
   });
 
   beforeEach(async () => {
+    await prisma.rating.deleteMany();
     await prisma.payment.deleteMany();
     await prisma.foodOrderItem.deleteMany();
     await prisma.foodOrder.deleteMany();