+++ src/modules/delivery/delivery.test.ts
@@ -18,6 +18,7 @@ describe('delivery module', () => {
   });
 
   beforeEach(async () => {
+    await prisma.rating.deleteMany();
     await prisma.foodOrderItem.deleteMany();
     await prisma.foodOrder.deleteMany();
     await prisma.payment.deleteMany();